requirements.txt
boto3==1.34.59
wheel 
setuptools
PyYAML
pyiceberg